+++ SimpleHttpServer/parse.h
@@ -15,4 +15,4 @@ extern const char *mimeTypeMap[][2];
 
 extern int GetMimeType(const char *file, size_t size);
 extern int ParseHttpRequestMessage(const char *message, HttpRequestMessage *structReq);
-extern int ParseHttpRequestUri(char *uri, char *resourcePath, char *args, int pathLen, int argsLen);
+extern int ParseHttpRequestUri(char *uri, char *resourcePath, char *args, size_t pathLen, size_t argsLen);
